+++ annotationengine/chunked_annotation.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, jsonify, abort
-from annotationengine.database import get_db
+from annotationengine.anno_database import get_db
 from flask import current_app, g
 import cloudvolume
